Fix: 삭제모달 수정
